(wip) This commit introduces the concept of PyQt-Workers.
This will eventually allow to start multiple downloads at the same time.

PyQt's QThread allows starting a UI-independent thread while still being
able to update the UI, if this is desired. So it's the best of both
worlds. The price to pay is complexity. And also a more tight coupling of the businees rules to the UI.